added str_c_string
returns a c-style string. might change the internals of this function,
currently it returns a pointer to the contents of the string, which
makes them modifyable by the end user. not optimal. but the other option
uses more memory and another alloc so idk which one would be better
quite frankly!